requirements.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.0
Django==4.1.5
django-rest-knox==4.2.0
djangorestframework==3.14.0
pycparser==2.21
pytz==2022.7.1
sqlparse==0.4.3
asgiref==3.5.1
Django==3.2.7
